+++ src/main/java/EdwinAlfaro_20210300/EdwinAlfaro_20210300/MODEL/DTO/LibroDTO.java
@@ -20,7 +20,7 @@ public class LibroDTO {
     private String isbn;
 
     @NotNull
-    private int año_publicacion;
+    private Long año_publicacion;
 
     private String genero;
